+++ gbdxtools/images/meta.py
@@ -6,36 +6,25 @@ from collections import Container, namedtuple
 import warnings
 import math
 
-from gbdxtools.ipe.io import to_geotiff
-from gbdxtools.ipe.util import RatPolyTransform, AffineTransform, pad_safe_positive, pad_safe_negative, IPE_TO_DTYPE, preview
+from gbdxtools.rda.io import to_geotiff
+from gbdxtools.rda.util import RatPolyTransform, AffineTransform, pad_safe_positive, pad_safe_negative, RDA_TO_DTYPE, preview
+from gbdxtools.images.mixins import PlotMixin, BandMethodsTemplate, Deprecations 
 
 from shapely import ops, wkt
 from shapely.geometry import box, shape, mapping, asShape
 from shapely.geometry.base import BaseGeometry
-try:
-    from rio_hist.match import histogram_match
-    has_rio = True
-except ImportError:
-    has_rio = False
-import mercantile
 
 import skimage.transform as tf
 
 import pyproj
 import dask
-from dask import sharedict, optimize
+from dask import sharedict, optimization
 from dask.delayed import delayed
 import dask.array as da
 from dask.base import is_dask_collection
 import numpy as np
 
 from affine import Affine
-
-try:
-    from matplotlib import pyplot as plt
-    has_pyplot = True
-except:
-    has_pyplot = False
 
 try:
     xrange
@@ -49,7 +38,7 @@ class DaskMeta(namedtuple("DaskMeta", ["dask", "name", "chunks", "dtype", "shape
     __slots__ = ()
     @classmethod
     def from_darray(cls, darr, new=tuple.__new__, len=len):
-        dsk, _ = optimize.cull(darr.dask, darr.__dask_keys__())
+        dsk, _ = optimization.cull(darr.dask, darr.__dask_keys__())
         itr = [dsk, darr.name, darr.chunks, darr.dtype, darr.shape]
         return cls._make(itr)
 
@@ -87,10 +76,8 @@ class DaskImage(da.Array):
     def read(self, bands=None, **kwargs):
         """
         Reads data from a dask array and returns the computed ndarray matching the given bands
-
         kwargs:
             bands (list): band indices to read from the image. Returns bands in the order specified in the list of bands.
-
         Returns:
             array (ndarray): a numpy array of image data
         """
@@ -102,10 +89,8 @@ class DaskImage(da.Array):
     def randwindow(self, window_shape):
         """
         Get a random window of a given shape from withing an image
-
         kwargs:
             window_shape (tuple): The desired shape of the returned image as (height, width) in pixels.
-
         Returns:
             image (dask): a new image object of the specified shape
         """
@@ -116,11 +101,9 @@ class DaskImage(da.Array):
     def iterwindows(self, count=64, window_shape=(256, 256)):
         """
         Iterate over random windows of an image
-
         kwargs:
             count (int): the number of the windows to generate. Defaults to 64, if `None` with continue to iterate over random windows until stopped.
             window_shape (tuple): The desired shape of each image as (height, width) in pixels.
-
         Returns:
             windows (generator): a generator of windows of the given shape
         """
@@ -131,121 +114,7 @@ class DaskImage(da.Array):
             for i in xrange(count):
                 yield self.randwindow(window_shape)
 
-# Mixin class that defines plotting methods and rgb/ndvi methods
-# used as a mixin to provide access to the plot method on
-# GeoDaskWrapper images and ipe images
-class PlotMixin(object):
-    @property
-    def _rgb_bands(self):
-        return [4, 2, 1]
-
-    @property
-    def _ndvi_bands(self):
-        if self.__class__.__name__ == 'Landsat8':
-            return [4,3]
-        elif self.__class__.__name__ == 'SENTINEL2':
-            return [1,3]
-        # elif self.__class__.__name__ == 'Sentinel2':
-        #     return [1, 3]
-        else:
-            return [1, 3]
-
-    @property
-    def _ndwi_bands(self):
-        if self.__class__.__name__ == 'Landsat8':
-            return [2, 4]
-        # elif self.__class__.__name__ == 'SENTINEL2':
-        #     return [2, 7]
-        else:
-            return [7, 0]
-
-    def ndwi(self):
-        """
-        Calculates Normalized Difference Water Index using Coastal and NIR2 bands for WV02, WV03.
-        For Landsat8 and sentinel2 calculated by using Green and NIR bands.
-
-        Returns: numpy array of ndwi values
-        """
-        data = self._read(self[self._ndwi_bands,...]).astype(np.float32)
-        return (data[1,:,:] - data[0,:,:]) / (data[0,:,:] + data[1,:,:])
-
-    def base_layer_match(self, blm=False, **kwargs):
-        rgb = self.rgb(**kwargs)
-        if not blm:
-            return rgb
-        from gbdxtools.images.tms_image import TmsImage
-        bounds = self._reproject(box(*self.bounds), from_proj=self.proj, to_proj="EPSG:4326").bounds
-        tms = TmsImage(zoom=self._calc_tms_zoom(self.affine[0]), bbox=bounds, **kwargs)
-        ref = np.rollaxis(tms.read(), 0, 3)
-        out = np.dstack([histogram_match(rgb[:,:,idx], ref[:,:,idx].astype(np.double)/255.0)
-                        for idx in xrange(rgb.shape[-1])])
-        return out
-
-    def rgb(self, **kwargs):
-        data = self._read(self[kwargs.get("bands", self._rgb_bands),...], **kwargs)
-        data = np.rollaxis(data.astype(np.float32), 0, 3)
-        lims = np.percentile(data, kwargs.get("stretch", [2, 98]), axis=(0, 1))
-        for x in xrange(len(data[0,0,:])):
-            top = lims[:,x][1]
-            bottom = lims[:,x][0]
-            data[:,:,x] = (data[:,:,x] - bottom) / float(top - bottom)
-        return np.clip(data, 0, 1)
-
-    def ndvi(self):
-        """
-        Calculates Normalized Difference Vegetation Index using NIR and Red of an image.
-
-        Returns: numpy array with ndvi values
-        """
-        data = self._read(self[self._ndvi_bands,...]).astype(np.float32)
-        return (data[0,:,:] - data[1,:,:]) / (data[0,:,:] + data[1,:,:])
-
-    def plot(self, spec="rgb", **kwargs):
-        if self.shape[0] == 1 or ("bands" in kwargs and len(kwargs["bands"]) == 1):
-            if "cmap" in kwargs:
-                cmap = kwargs["cmap"]
-                del kwargs["cmap"]
-            else:
-                cmap = "Greys_r"
-            self._plot(tfm=self._single_band, cmap=cmap, **kwargs)
-        else:
-            if spec == "rgb" and self._has_token(**kwargs):
-                self._plot(tfm=self.base_layer_match, **kwargs)
-            else:
-                self._plot(tfm=getattr(self, spec), **kwargs)
-
-    def _has_token(self, **kwargs):
-        if "access_token" in kwargs or "MAPBOX_API_KEY" in os.environ:
-            return True
-        else:
-            return False
-
-    def _plot(self, tfm=lambda x: x, **kwargs):
-        assert has_pyplot, "To plot images please install matplotlib"
-        assert self.shape[1] and self.shape[-1], "No data to plot, dimensions are invalid {}".format(str(self.shape))
-
-        f, ax1 = plt.subplots(1, figsize=(kwargs.get("w", 10), kwargs.get("h", 10)))
-        ax1.axis('off')
-        plt.imshow(tfm(**kwargs), interpolation='nearest', cmap=kwargs.get("cmap", None))
-        plt.show(block=False)
-
-    def _read(self, data, **kwargs):
-        if hasattr(data, 'read'):
-            return data.read(**kwargs)
-        else:
-            return data.compute(get=threaded_get)
-
-    def _single_band(self, **kwargs):
-        return self._read(self[0,:,:], **kwargs)
-
-    def _calc_tms_zoom(self, scale):
-        for z in range(15,20):
-            b = mercantile.bounds(0,0,z)
-            if scale > math.sqrt((b.north - b.south)*(b.east - b.west) / (256*256)):
-                return z
-
-
-class GeoDaskImage(DaskImage, Container, PlotMixin):
+class GeoDaskImage(DaskImage, Container, PlotMixin, BandMethodsTemplate, Deprecations):
     _default_proj = "EPSG:4326"
 
     def map_blocks(self, *args, **kwargs):
@@ -259,7 +128,6 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
     @property
     def affine(self):
         """ The geo transform of the image
-
         Returns:
             affine (dict): The image's affine transform
         """
@@ -269,7 +137,6 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
     @property
     def bounds(self):
         """ Access the spatial bounding box of the image
-
         Returns:
             bounds (list): list of bounds in image projected coordinates (minx, miny, maxx, maxy)
         """
@@ -282,12 +149,10 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
 
     def aoi(self, **kwargs):
         """ Subsets the Image by the given bounds
-
         kwargs:
             bbox: optional. A bounding box array [minx, miny, maxx, maxy]
             wkt: optional. A WKT geometry string
             geojson: optional. A GeoJSON geometry dictionary
-
         Returns:
             image (ndarray): an image instance
         """
@@ -299,11 +164,9 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
 
     def pxbounds(self, geom, clip=False):
         """ Returns the bounds of a geometry object in pixel coordinates
-
         args:
             geom: Shapely geometry object or GeoJSON as Python dictionary or WKT string
             clip (bool): Clip the bounds to the min/max extent of the image
-
         Returns:
             list of bounds in pixels [min x, min y, max x, max y] clipped to image bounds
         """
@@ -337,13 +200,11 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
 
     def geotiff(self, **kwargs):
         """ Creates a geotiff on the filesystem
-
         kwargs:
             path (str): optional. The path to save the geotiff to.
             bands (list): optional. A list of band indices to save to the output geotiff ([4,2,1])
             dtype (str): optional. The data type to assign the geotiff to ("float32", "uint16", etc)
             proj (str): optional. An EPSG proj string to project the image data into ("EPSG:32612")
-
         Returns:
             path (str): the path to created geotiff
         """
@@ -358,16 +219,14 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
         """
         Delayed warp across an entire AOI or Image
         creates a new dask image by deferring calls to the warp_geometry on chunks
-
         kwargs:
             dem (ndarray): optional. A DEM for warping to specific elevation planes
             proj (str): optional. An EPSG proj string to project the image data into ("EPSG:32612")
-
         Returns:
             image (dask): a warped image as deferred image array (a dask)
         """
         try:
-            img_md = self.ipe.metadata["image"]
+            img_md = self.rda.metadata["image"]
             x_size = img_md["tileXSize"]
             y_size = img_md["tileYSize"]
         except (AttributeError, KeyError):
@@ -381,12 +240,12 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
             from_proj = self.proj
 
         try:
-            # NOTE: this only works on images that have IPE rpcs metadata
-            center = wkt.loads(self.ipe.metadata["image"]["imageBoundsWGS84"]).centroid
-            g = box(*(center.buffer(self.ipe.metadata["rpcs"]["gsd"] / 2).bounds))
+            # NOTE: this only works on images that have rda rpcs metadata
+            center = wkt.loads(self.rda.metadata["image"]["imageBoundsWGS84"]).centroid
+            g = box(*(center.buffer(self.rda.metadata["rpcs"]["gsd"] / 2).bounds))
             tfm = partial(pyproj.transform, pyproj.Proj(init="EPSG:4326"), pyproj.Proj(init=proj))
             gsd = kwargs.get("gsd", ops.transform(tfm, g).area ** 0.5)
-            current_bounds = wkt.loads(self.ipe.metadata["image"]["imageBoundsWGS84"]).bounds
+            current_bounds = wkt.loads(self.rda.metadata["image"]["imageBoundsWGS84"]).bounds
         except (AttributeError, KeyError, TypeError):
             tfm = partial(pyproj.transform, pyproj.Proj(init=self.proj), pyproj.Proj(init=proj))
             gsd = kwargs.get("gsd", (ops.transform(tfm, shape(self)).area / (self.shape[1] * self.shape[2])) ** 0.5 )
@@ -405,7 +264,7 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
         num_bands = self.shape[0]
 
         try:
-            dtype = IPE_TO_DTYPE[img_md["dataType"]]
+            dtype = RDA_TO_DTYPE[img_md["dataType"]]
         except:
             dtype = 'uint8'
 
@@ -437,7 +296,7 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
                 ymin = y * y_size
                 geometry = px_to_geom(xmin, ymin)
                 daskmeta["dask"][(daskmeta["name"], 0, y, x)] = (self._warp, geometry, gsd, dem, proj, dtype, 5)
-        daskmeta["dask"], _ = optimize.cull(sharedict.merge(daskmeta["dask"], *dasks), list(daskmeta["dask"].keys()))
+        daskmeta["dask"], _ = optimization.cull(sharedict.merge(daskmeta["dask"], *dasks), list(daskmeta["dask"].keys()))
 
         gi = mapping(full_bounds)
         gt = AffineTransform(gtf, proj)
@@ -567,7 +426,10 @@ class GeoDaskImage(DaskImage, Container, PlotMixin):
                     return self[arg0, :, :]
 
             elif len(geometry) == 3:
-                nbands, ysize, xsize = self.shape
+                try:
+                    nbands, ysize, xsize = self.shape
+                except:
+                    ysize, xsize = self.shape
                 band_idx, y_idx, x_idx = geometry
                 if y_idx == Ellipsis:
                     y_idx = slice(0, ysize)